--- lucet-validate/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "lucet-validate"
-version = "0.7.0-dev"
-description = "Parse and validate webassembly files against witx interface"
-homepage = "https://github.com/fastly/lucet"
-repository = "https://github.com/fastly/lucet"
-license = "Apache-2.0 WITH LLVM-exception"
-categories = ["wasm"]
-authors = ["Pat Hickey <phickey@fastly.com>"]
-edition = "2018"
-
-[lib]
-crate-type=["rlib"]
-
-[[bin]]
-name = "lucet-validate"
-path = "src/main.rs"
-
-[dependencies]
-clap = "2"
-witx = { path = "../wasmtime/crates/wasi-common/WASI/tools/witx", version = "0.8.5" }
-cranelift-entity = { path = "../wasmtime/cranelift/entity", version = "0.65.0" }
-thiserror = "1.0.4"
-wasmparser = "0.52.0"
-
-[dev-dependencies]
-lucet-wasi = { path = "../lucet-wasi", version = "=0.7.0-dev" }
-lucet-wasi-sdk = { path = "../lucet-wasi-sdk", version = "=0.7.0-dev" }
-tempfile = "3.0"
-wabt = "0.9.2"
