+++ lucetc/Cargo.toml
@@ -25,8 +25,8 @@ cranelift-object = { path = "../wasmtime/cranelift/object", version = "0.65.0" }
 cranelift-wasm = { path = "../wasmtime/cranelift/wasm", version = "0.65.0" }
 target-lexicon = "0.10"
 lucet-module = { path = "../lucet-module", version = "=0.7.0-dev" }
-lucet-validate = { path = "../lucet-validate", version = "=0.7.0-dev" }
 lucet-wiggle-generate = { path = "../lucet-wiggle/generate", version = "=0.7.0-dev" }
+witx = { path = "../wasmtime/crates/wasi-common/WASI/tools/witx", version = "0.8.5" }
 wasmparser = "0.58.0"
 clap="2.32"
 log = "0.4"
