+++ lucetc/src/error.rs
@@ -1,4 +1,5 @@
 use crate::types::SignatureError;
+use crate::validate::Error as ValidationError;
 use cranelift_module::ModuleError as ClifModuleError;
 use cranelift_wasm::WasmError as ClifWasmError;
 use lucet_module::error::Error as LucetModuleError;
@@ -15,7 +16,7 @@ pub enum Error {
     #[error("Lucet Module: {0}")]
     LucetModule(#[from] LucetModuleError),
     #[error("Lucet validation: {0}")]
-    LucetValidation(#[from] lucet_validate::Error),
+    LucetValidation(#[from] ValidationError),
     #[error("I/O: {0}")]
     IOError(#[from] std::io::Error),
     #[error("Converting to Wasm signature: {0}")]
